+++ Restaurants.Application/Restuarants/Commands/DeleteRestaurant/DeleteRestaurantHandler.cs
@@ -1,7 +1,9 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories.Restaurants;
 using System;
 using System.Collections.Generic;
@@ -12,7 +14,8 @@ using System.Threading.Tasks;
 namespace Restaurants.Application.Restuarants.Commands.DeleteRestaurant
 {
     public class DeleteRestaurantHandler(ILogger<DeleteRestaurantHandler> logger,
-        IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+        IRestaurantsRepository restaurantsRepository,
+        IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
     {
         public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
         {
@@ -20,6 +23,9 @@ namespace Restaurants.Application.Restuarants.Commands.DeleteRestaurant
             var restaurant = await restaurantsRepository.GetRestaurantByIdAsync(request.Id);
             if (restaurant is null)
                 throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+
+            if (!restaurantAuthorizationService.Authorized(restaurant, ResourceOperation.Delete))
+                throw new ForbidException();
 
             await restaurantsRepository.DeleteAsync(restaurant);
         }
